requirements.txt
peewee==3.17.0
aiogram==3.4.1
python-i18n[YAML]==0.3.9
loguru==0.7.2
redis==5.0.1
psycopg2-binary
python-dotenv==1.0.0
requests==2.31.0
aiosseclient==0.1.1
base58==2.1.1
ruamel.yaml==0.18.5